backend/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
python-dotenv
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
twilio
